+++ tests/dummy/app/app.js
@@ -3,7 +3,6 @@ import Resolver from './resolver';
 import loadInitializers from 'ember-load-initializers';
 import config from './config/environment';
 import './default-user-settings/index';
-import './models/custom-inflector-rules';
 
 const App = Application.extend({
   modulePrefix: config.modulePrefix,